Fixed obsolescent calls in tests and some controllers.
Fixed possible bug in Voucher#not_already_used: initial check was for
voucher.showdate.nil?, which actually COULD be true even if
voucher.showdate_id was nonzero.  This only happens in the case where
voucher.showdate_id is specifically set after the model instance has
been loaded and without saying voucher << showdate, so it was only
manifesting in the testcase code but would not manifest normally since
there are no codepaths where this is done. Still, it's now fixed and
not_already_used checks the showdate_id of the voucher first.  

Haml'ized most of the remaining templates.  

Added code to deploy.rb to allow deploying from a branch.