New popup menu on Vouchertypes screen allows you to show subscriptions   only, all bundles, or all vouchertypes.
There used to be a "Whoops" error that occurred when adding a Voucher
Type to All Dates when one of those dates *already* allowed that Voucher
Type.  This has been fixed; any such dates are now ignored (and a
message is printed to that effect), while the Voucher Type is correctly
added to the remaining dates.